AcctIdx: StoredSize is u32

+++ runtime/src/account_info.rs
@@ -11,7 +11,8 @@ use crate::{
 pub type Offset = usize;
 
 /// bytes used to store this account in append vec
-pub type StoredSize = usize;
+/// Note this max needs to be big enough to handle max data len of 10MB, which is a const
+pub type StoredSize = u32;
 
 /// specify where account data is located
 #[derive(Debug)]
